Add files via upload
This is a beta version of the gui.

- Working features are: Add files work, Start Encoding works with the proggress bar. Remove selected wot´rks if no encode is in progress. Allencoder are run parallel so be carefull may hog all of the available cpu power!!!

- Not Working: The Stop Encoding button does nothing for now

Default encoding is set to be h264_nvenc for my hardware and speed reuirements reasons. This is still pretty slow tho, but the quality is great for the achieved compression. Hevc could improve file size if legacy codecs are not needed. I shold add preset list menu sometime but i am pretty lazy...

Builds fine with vs2022 community